modification2

+++ src/main/java/tn/m104/rh/entity/Student.java
@@ -19,6 +19,6 @@ public class Student {
     Integer rollNumber;
     String name;
     String address;
-    Double percentage;
+    Integer percentage;
 
 }
